Moving generated graphql models to /gen/graphql.ts and adding .gitignore
