+++ .gitignore
@@ -72,6 +72,7 @@ schema.ts
 possibleTypes.json
 fragmentTypes.json
 **/gen/fetch/**
+**/gen/graphql.ts
 
 # IDE - Visual Studio
 *.suo
