+++ libs/shared/components/codegen.yml
@@ -3,7 +3,7 @@ schema:
 documents:
   - libs/shared/components/src/**/*.graphql.ts
 generates:
-  libs/shared/components/src/schema.ts:
+  libs/shared/components/gen/graphql.ts:
     plugins:
       - typescript
       - typescript-operations
